# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: MoveOps API
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /auth/login:
    post:
      operationId: PostAuthLogin
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/logout:
    post:
      operationId: PostAuthLogout
      summary: Logout current session
      responses:
        '204':
          description: Logged out
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/me:
    get:
      operationId: GetAuthMe
      summary: Get current user and tenant
      responses:
        '200':
          description: Current session principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/csrf:
    get:
      operationId: GetAuthCsrf
      summary: Get csrf token for session
      responses:
        '200':
          description: CSRF token response
          content:
            application/json:
              schema:
                type: object
                required: [csrfToken]
                properties:
                  csrfToken:
                    type: string
        default:
          $ref: '#/components/responses/ErrorResponse'
  /customers:
    post:
      operationId: PostCustomers
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /customers/{customerId}:
    get:
      operationId: GetCustomersCustomerId
      summary: Get customer by id
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /estimates:
    get:
      operationId: GetEstimates
      summary: List estimates (paged)
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [draft, converted]
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: cursor
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Estimates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateListResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: PostEstimates
      summary: Create a draft estimate
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEstimateRequest'
      responses:
        '201':
          description: Draft estimate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateResponse'
        '200':
          description: Idempotent replay returned existing estimate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /estimates/{estimateId}:
    get:
      operationId: GetEstimatesEstimateId
      summary: Get estimate by id
      parameters:
        - in: path
          name: estimateId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Estimate details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: PatchEstimatesEstimateId
      summary: Update estimate fields
      parameters:
        - in: path
          name: estimateId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEstimateRequest'
      responses:
        '200':
          description: Estimate updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /estimates/{estimateId}/convert:
    post:
      operationId: PostEstimatesEstimateIdConvert
      summary: Convert estimate to job (idempotent)
      parameters:
        - in: path
          name: estimateId
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Job created from estimate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '200':
          description: Idempotent replay returned existing job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /estimates/{estimateId}/inventory:
    get:
      operationId: GetEstimatesEstimateIdInventory
      summary: Get estimate inventory
      parameters:
        - in: path
          name: estimateId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Estimate inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateInventoryResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      operationId: PutEstimatesEstimateIdInventory
      summary: Replace estimate inventory and recalculate totals
      parameters:
        - in: path
          name: estimateId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceEstimateInventoryRequest'
      responses:
        '200':
          description: Inventory updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateInventoryResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /estimates/{estimateId}/inventory-share-links:
    post:
      operationId: PostEstimatesEstimateIdInventoryShareLinks
      summary: Create customer inventory share link and send email
      parameters:
        - in: path
          name: estimateId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryShareLinkRequest'
      responses:
        '201':
          description: Share link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInventoryShareLinkResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /public/inventory/{token}:
    get:
      operationId: GetPublicInventoryToken
      summary: Fetch customer inventory payload via share token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public inventory payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicInventoryResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      operationId: PutPublicInventoryToken
      summary: Submit customer inventory payload via share token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceEstimateInventoryRequest'
      responses:
        '200':
          description: Public inventory updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicInventoryResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /calendar:
    get:
      operationId: GetCalendar
      summary: List scheduled jobs for a monthly calendar range
      description: >
        Returns jobs where `scheduledDate` is within `[from, to)` (inclusive `from`, exclusive `to`).
        `userId` and `departmentId` are accepted as filter placeholders but are currently ignored.
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: phase
          required: false
          schema:
            type: string
            enum: [booked, scheduled, completed, cancelled]
        - in: query
          name: jobType
          required: false
          schema:
            type: string
            enum: [local, long_distance, other]
        - in: query
          name: userId
          required: false
          description: Placeholder filter; accepted but currently ignored.
          schema:
            type: string
            format: uuid
        - in: query
          name: departmentId
          required: false
          description: Placeholder filter; accepted but currently ignored.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Calendar jobs in range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /dashboard/summary:
    get:
      operationId: GetDashboardSummary
      summary: Dashboard summary counts for the current tenant
      responses:
        '200':
          description: Summary counts (permission-aware)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummaryResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /jobs/{jobId}:
    get:
      operationId: GetJobsJobId
      summary: Get job by id
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: PatchJobsJobId
      summary: Update minimal job scheduling fields
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobRequest'
      responses:
        '200':
          description: Job updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /exports/customers.csv:
    get:
      operationId: GetExportsCustomersCsv
      summary: Export tenant customers CSV
      responses:
        '200':
          description: Customers CSV
          content:
            text/csv:
              schema:
                type: string
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthSessionResponse:
      type: object
      required: [user, tenant]
      properties:
        user:
          $ref: '#/components/schemas/User'
        tenant:
          $ref: '#/components/schemas/Tenant'
...
    CreateCustomerRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
    Customer:
      type: object
      required: [id, tenantId, firstName, lastName, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...